requirements.txt
fastapi
pydantic
uvicorn
jinja2
pymongo
Werkzeug
python-dotenv
fastapi-jwt-auth